L&C python/client/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
rich==13.7.0
pytest==8.0.0
pytest-mock==3.12.0 